feat: add service mesh policies to coordinator relations (#90)
* feat: add service mesh policy management for metrics-endpoint

Previously, the Coordinator did not request AuthorizationPolicies to enable anyone related via the metrics-endpoint to talk to Coordinator.  This meant that, in a hardened mesh, the metrics scraper would not have permission to scrape metrics.

This commit fixes the issue by adding Policies to cover this.  Specifically, we:
* Add a UnitPolicy to allow metrics scraping of the Coordinator's nginx pod
* Add a AppPolicy to allow metrics scraping of the Worker pods via the Coordinator's metrics proxy
* change the worker tester container image to a basic image that offers a /metrics endpoint to allow for testing metrics collection via coordinator's proxy

Driveby changes included here:
* fix flakey test_configure_service_mesh by adding retry mechanics
* expose the NginxPrometheusExporter's metrics port via a property